Moving softmax to its own plugin
- first steps towards adding a Keras support plugin
