+++ phylanx/plugins/keras_support/keras_support.hpp
@@ -1,0 +1,11 @@
+//  Copyright (c) 2019 Hartmut Kaiser
+//
+//  Distributed under the Boost Software License, Version 1.0. (See accompanying
+//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#if !defined(PHYLANX_PLUGINS_KERAS_SUPPORT_MAR_11_2019_0441PM)
+#define PHYLANX_PLUGINS_KERAS_SUPPORT_MAR_11_2019_0441PM
+
+#include <phylanx/plugins/keras_support/softmax_operation.hpp>
+
+#endif
