+++ phylanx/plugins/matrixops/matrixops.hpp
@@ -36,7 +36,6 @@
 #include <phylanx/plugins/matrixops/reshape_operation.hpp>
 #include <phylanx/plugins/matrixops/shuffle_operation.hpp>
 #include <phylanx/plugins/matrixops/slicing_operation.hpp>
-#include <phylanx/plugins/matrixops/softmax_operation.hpp>
 #include <phylanx/plugins/matrixops/squeeze_operation.hpp>
 #include <phylanx/plugins/matrixops/stack_operation.hpp>
 #include <phylanx/plugins/matrixops/tile_operation.hpp>
