+++ src/plugins/matrixops/matrixops.cpp
@@ -95,8 +95,6 @@ PHYLANX_REGISTER_PLUGIN_FACTORY(shuffle_operation_plugin,
     phylanx::execution_tree::primitives::shuffle_operation::match_data);
 PHYLANX_REGISTER_PLUGIN_FACTORY(slicing_operation_plugin,
     phylanx::execution_tree::primitives::slicing_operation::match_data[0]);
-PHYLANX_REGISTER_PLUGIN_FACTORY(softmax_operation_plugin,
-    phylanx::execution_tree::primitives::softmax_operation::match_data);
 PHYLANX_REGISTER_PLUGIN_FACTORY(squeeze_operation_plugin,
     phylanx::execution_tree::primitives::squeeze_operation::match_data);
 PHYLANX_REGISTER_PLUGIN_FACTORY(stack_operation_plugin,
