+++ tests/unit/plugins/matrixops/CMakeLists.txt
@@ -39,7 +39,6 @@ set(tests
     row_slicing
     shuffle_operation
     slicing_operation
-    softmax_operation
     squeeze_operation
     tile_operation
     transpose_operation
